Split class game over, and general changes
Split function gameOver in app to own file GameOver.ts. If the player lose against the Computer a lose message is shown. Computer vs Computer win message include move count.
General change, fixes and comments.
