+++ typescript/classes/Board.ts
@@ -1,4 +1,3 @@
-// Board.ts
 export default class Board {
 
   matrix: string[][];
